Refactor BooksPage component
