--- src/pages/BooksPage.js
@@ -1,5 +1,0 @@
-const BooksPage = () => (
-  <h1> Books </h1>
-);
-
-export default BooksPage;
